websocket repair
