+++ hw15-messageSystem/src/main/java/ru/otus/config/WebSocketConfig.java
@@ -19,6 +19,6 @@ public class WebSocketConfig implements WebSocketMessageBrokerConfigurer {
 
     @Override
     public void registerStompEndpoints(StompEndpointRegistry registry) {
-        registry.addEndpoint("/websocket").withSockJS();
+        registry.addEndpoint("/user-websocket").withSockJS();
     }
 }
