+++ hw15-messageSystem/src/main/resources/templates/add_user.html
@@ -23,7 +23,8 @@
             <input type="password" id="password" class="form-control" placeholder="******">
         </div>
         <br/>
-        <button id="send" class="btn btn-default" type="submit">Добавить пользователя</button>
+        <input type="button" onclick="createUser()" value="Добавить пользователя">
+<!--        <button id="send" class="btn btn-default" type="submit">Добавить пользователя</button>-->
     </form>
     <table id="conversation" class="table table-striped">
         <thead>
